# custom_gpt/openapi.yaml
openapi: 3.1.0
info:
  title: K-Trader Read-only API
  version: phase10-production-v2
  description: >-
    Read-only K-Trader Action API. MT4 market context is exposed to ChatGPT only
    through bounded compact endpoints; the full canonical MT4 context remains a
    backend/E2E route and is intentionally omitted from this Action schema.
servers:
  - url: https://ktrader-api.duckdns.org
    description: Accepted production HTTPS origin.
security:
  - BearerAuth: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Check K-Trader API and scanner readiness
      security: []
      responses:
        '200':
          description: API/scanner health state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Health'}
  /v1/scanner/status:
    get:
      operationId: getScannerStatus
      summary: Get scanner runtime status and data readiness
      responses:
        '200':
          description: Scanner runtime status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScannerStatus'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /v1/universe:
    get:
      operationId: listUniverse
      summary: List the current ranked tradable universe
      parameters:
        - name: provider_id
          in: query
          required: false
          schema: {type: string}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: Ranked universe
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UniverseList'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /v1/market/{symbol}:
    get:
      operationId: getMarketSnapshot
      summary: Get the latest normalized scanner market snapshot
      parameters:
        - name: symbol
          in: path
          required: true
          schema: {type: string}
        - name: provider_id
          in: query
          required: false
          schema: {type: string}
      responses:
        '200':
          description: Scanner market snapshot
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Ambiguous'}
  /v1/mt4/market-context-summary/{symbol}:
    get:
      operationId: getMT4MarketContextSummary
      summary: Get compact fresh validated MT4 market context without candle arrays
      description: >-
        Preferred MT4 entry point for ChatGPT. Returns market facts plus D1/H1/M15/M5
        scope metadata only. BrokerServer timestamps are opaque wall-clock strings and
        must not be interpreted as UTC.
      parameters:
        - name: symbol
          in: path
          required: true
          description: MT4 broker symbol, for example USDTRY.
          schema: {type: string}
        - name: market
          in: query
          required: false
          schema: {type: string, default: forex}
      responses:
        '200':
          description: Compact schema 1.1 MT4 context
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MT4MarketContextSummary'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {description: Invalid symbol or market request}
        '503': {description: MT4 market context disabled, busy, or temporarily unavailable}
  /v1/mt4/candles/{symbol}:
    get:
      operationId: getMT4Candles
      summary: Get a bounded tail of one validated MT4 timeframe
      description: >-
        Use after getMT4MarketContextSummary. Returns at most 120 closed candles from
        exactly one of D1, H1, M15, M5 to stay within ChatGPT Action response limits.
      parameters:
        - name: symbol
          in: path
          required: true
          schema: {type: string}
        - name: market
          in: query
          required: false
          schema: {type: string, default: forex}
        - name: timeframe
          in: query
          required: false
          schema: {type: string, enum: [D1, H1, M15, M5], default: M5}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 120, default: 60}
      responses:
        '200':
          description: Bounded closed MT4 candles
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MT4CandleSeries'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {description: Invalid symbol, market, timeframe, or limit}
        '503': {description: MT4 market context disabled, busy, or temporarily unavailable}
  /v1/candles/{symbol}:
    get:
      operationId: getCandles
      summary: Get confirmed normalized scanner candles
      parameters:
        - name: symbol
          in: path
          required: true
          schema: {type: string}
        - name: provider_id
          in: query
          required: false
          schema: {type: string}
        - name: interval
          in: query
          required: false
          schema: {type: string, enum: [5m, 15m, 1h, 4h, 1d], default: 5m}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 500, default: 100}
      responses:
        '200':
          description: Scanner candle series
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Ambiguous'}
  /v1/analysis/{symbol}:
    get:
      operationId: getAnalysis
      summary: Get the current Trading Engine decision for a scanner symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema: {type: string}
        - name: provider_id
          in: query
          required: false
          schema: {type: string}
      responses:
        '200':
          description: Trading decision
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Ambiguous'}
  /v1/candidates:
    get:
      operationId: listCandidates
      summary: List ranked current setup candidates including NO_TRADE
      parameters:
        - name: provider_id
          in: query
          required: false
          schema: {type: string}
        - name: grade
          in: query
          required: false
          schema: {type: string, enum: [A+, A, B, C]}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: Candidate decisions
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /v1/signals:
    get:
      operationId: listSignals
      summary: List currently tradable A or A+ scanner signals
      parameters:
        - name: provider_id
          in: query
          required: false
          schema: {type: string}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: Tradable scanner signals
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        '401': {$ref: '#/components/responses/Unauthorized'}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Configure with the production KTRADER_ACTION_API_KEY value.
  responses:
    Unauthorized: {description: Missing or invalid Bearer API key}
    NotFound: {description: Requested scanner symbol or analysis is not available}
    Ambiguous: {description: Symbol exists on multiple scanner providers; repeat with provider_id}
  schemas:
    Health:
      type: object
      required: [status, mode, api_version, data_ready, scanner_status, action_auth_enabled, generated_at]
      properties:
        status: {type: string}
        mode: {type: string, enum: [read_only]}
        api_version: {type: string}
        data_ready: {type: boolean}
        scanner_status: {type: string}
        provider_id: {type: [string, 'null']}
        action_auth_enabled: {type: boolean}
        generated_at: {type: string, format: date-time}
    ScannerStatus:
      type: object
      properties:
        status: {type: string}
        provider_id: {type: [string, 'null']}
        universe_size: {type: integer}
        data_ready: {type: boolean}
        last_scan_at: {type: [string, 'null'], format: date-time}
        last_error: {type: [string, 'null']}
        cycle_id: {type: integer}
        symbols_ready: {type: integer}
        symbols_failed: {type: integer}
        live_streaming: {type: boolean}
        last_cycle_duration_seconds: {type: [number, 'null']}
        started_at: {type: string, format: date-time}
    UniverseList:
      type: object
      properties:
        count: {type: integer}
        provider_id: {type: [string, 'null']}
        items: {type: array, items: {type: object, additionalProperties: true}}
    MT4Base:
      type: object
      properties:
        provider_id: {type: string, enum: [kai_mt4]}
        source_name: {type: string, enum: [MT4]}
        symbol: {type: string}
        market: {type: string}
        schema_version: {type: string, enum: ['1.1']}
        run_id: {type: string}
        as_of: {type: [string, 'null'], description: Opaque BrokerServer wall-clock timestamp; not UTC.}
        last_tick_time: {type: [string, 'null'], description: Opaque BrokerServer wall-clock timestamp; not UTC.}
        time_source: {type: string, enum: [BrokerServer]}
        timestamp_semantics: {type: string, enum: [BROKER_SERVER_WALL_CLOCK_OPAQUE]}
        utc_offset_minutes: {type: 'null'}
        closed_bars_only: {type: boolean, enum: [true]}
    MT4MarketContextSummary:
      allOf:
        - $ref: '#/components/schemas/MT4Base'
        - type: object
          required: [provider_id, source_name, symbol, market, schema_version, run_id, closed_bars_only, bid, ask, spread, digits, point, scopes]
          properties:
            bid: {type: number}
            ask: {type: number}
            spread: {type: number}
            digits: {type: integer}
            point: {type: number}
            tick_value: {type: [number, 'null']}
            tick_size: {type: [number, 'null']}
            contract_size: {type: [number, 'null']}
            min_lot: {type: [number, 'null']}
            max_lot: {type: [number, 'null']}
            lot_step: {type: [number, 'null']}
            stop_level: {type: [number, 'null']}
            freeze_level: {type: [number, 'null']}
            trade_allowed: {type: [boolean, 'null']}
            terminal_connected: {type: [boolean, 'null']}
            market_open: {type: [boolean, 'null']}
            swap_long: {type: [number, 'null']}
            swap_short: {type: [number, 'null']}
            scopes:
              type: object
              required: [D1, H1, M15, M5]
              properties:
                D1: {$ref: '#/components/schemas/MT4ScopeSummary'}
                H1: {$ref: '#/components/schemas/MT4ScopeSummary'}
                M15: {$ref: '#/components/schemas/MT4ScopeSummary'}
                M5: {$ref: '#/components/schemas/MT4ScopeSummary'}
              additionalProperties: false
    MT4ScopeSummary:
      type: object
      properties:
        timeframe: {type: string, enum: [D1, H1, M15, M5]}
        snapshot_id: {type: [string, 'null']}
        bar_depth: {type: [integer, 'null']}
        bars_available: {type: integer}
        latest_closed_bar_time: {type: [string, 'null'], description: Opaque BrokerServer wall-clock timestamp; not UTC.}
        current_bar_time: {type: [string, 'null'], description: Opaque BrokerServer wall-clock timestamp; not UTC.}
    MT4CandleSeries:
      allOf:
        - $ref: '#/components/schemas/MT4Base'
        - type: object
          required: [provider_id, source_name, symbol, market, schema_version, run_id, closed_bars_only, timeframe, bar_depth, count, bars]
          properties:
            timeframe: {type: string, enum: [D1, H1, M15, M5]}
            snapshot_id: {type: [string, 'null']}
            bar_depth: {type: integer}
            latest_closed_bar_time: {type: string, description: Opaque BrokerServer wall-clock timestamp; not UTC.}
            current_bar_time: {type: string, description: Opaque BrokerServer wall-clock timestamp; not UTC.}
            count: {type: integer, minimum: 0, maximum: 120}
            bars:
              type: array
              maxItems: 120
              items: {$ref: '#/components/schemas/MT4Bar'}
    MT4Bar:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time: {type: string, description: Opaque BrokerServer wall-clock timestamp; not UTC.}
        open: {type: number}
        high: {type: number}
        low: {type: number}
        close: {type: number}
        volume: {type: number}